[WIP] Modifications that help make having a ping queue easier. A lot of cleanup needed, especially around error handling, and timeouts.
Also, this method doesn't fully accomplish what I really wanted, but I'm still thinking about the best way to accomplish that. The need to call connect on the socket hurts reuse of the socket, which is part of the benefit of UDP. That said, to accomplish using at most 2 sockets I'd need to move all the socket send/recv code to a separate layer (be it PingQueue or another class to act as the "pipe") and then associate the response with the correct ping based on the recvfrom socket address that gets returned.